requirements.txt
requests>=2.32,<3
python-dotenv>=1.0,<2
Pillow>=10,<12
